requirements.txt
asgiref==3.2.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-dotenv==1.4.2
django-extensions==2.2.8
django-rest-framework==0.1.0
django-stubs==1.2.0
djangorestframework==3.11.0
drf-yasg==1.17.0
gunicorn==20.0.4
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
mypy==0.740
mypy-extensions==0.4.3
packaging==20.1
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sentry-sdk==0.14.1
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.22
uuid==1.30
wcwidth==0.1.8
